Fix archives/zip tests on Windows
