+++ helpers/archives/zip_create_test.go
@@ -5,7 +5,7 @@ import (
 	"bytes"
 	"io/ioutil"
 	"os"
-	"syscall"
+	"runtime"
 	"testing"
 
 	"github.com/sirupsen/logrus"
@@ -55,17 +55,6 @@ func createTestDirectory(t *testing.T, csb charsetByte) string {
 	return name
 }
 
-func createTestPipe(t *testing.T, csb charsetByte) string {
-	name := "test_pipe"
-	if csb == multiBytes {
-		name = "テストパイプ"
-	}
-
-	err := syscall.Mkfifo(name, 0600)
-	assert.NoError(t, err)
-	return name
-}
-
 func createTestGitPathFile(t *testing.T, csb charsetByte) string {
 	_, err := os.Stat(".git")
 	if err != nil {
@@ -109,12 +98,25 @@ func TestZipCreate(t *testing.T) {
 			createTestFile(t, singleByte),
 			createSymlinkFile(t, singleByte),
 			createTestDirectory(t, singleByte),
-			createTestPipe(t, singleByte),
 			createTestFile(t, multiBytes),
 			createSymlinkFile(t, multiBytes),
 			createTestDirectory(t, multiBytes),
-			createTestPipe(t, multiBytes),
 			"non_existing_file.txt",
+		}
+
+		// only check how pipes are handled on unix
+		if runtime.GOOS != "windows" {
+			paths = append(
+				paths,
+				createTestPipe(t, singleByte),
+				createTestPipe(t, multiBytes),
+			)
+		}
+
+		expectedMode := os.FileMode(0640)
+		if runtime.GOOS == "windows" {
+			// windows doesn't support the same permissions as Linux
+			expectedMode = 0666
 		}
 
 		f, err := os.Create(fileName)
@@ -132,7 +134,7 @@ func TestZipCreate(t *testing.T) {
 		assert.Len(t, archive.File, 6)
 
 		assert.Equal(t, paths[0], archive.File[0].Name)
-		assert.Equal(t, os.FileMode(0640), archive.File[0].Mode().Perm())
+		assert.Equal(t, expectedMode, archive.File[0].Mode().Perm())
 		assert.NotEmpty(t, archive.File[0].Extra)
 
 		assert.Equal(t, paths[1], archive.File[1].Name)
@@ -141,13 +143,13 @@ func TestZipCreate(t *testing.T) {
 		assert.NotEmpty(t, archive.File[2].Extra)
 		assert.True(t, archive.File[2].Mode().IsDir())
 
-		assert.Equal(t, paths[4], archive.File[3].Name)
-		assert.Equal(t, os.FileMode(0640), archive.File[3].Mode().Perm())
+		assert.Equal(t, paths[3], archive.File[3].Name)
+		assert.Equal(t, expectedMode, archive.File[3].Mode().Perm())
 		assert.NotEmpty(t, archive.File[3].Extra)
 
-		assert.Equal(t, paths[5], archive.File[4].Name)
+		assert.Equal(t, paths[4], archive.File[4].Name)
 
-		assert.Equal(t, paths[6]+"/", archive.File[5].Name)
+		assert.Equal(t, paths[5]+"/", archive.File[5].Name)
 		assert.NotEmpty(t, archive.File[5].Extra)
 		assert.True(t, archive.File[5].Mode().IsDir())
 	})
@@ -163,6 +165,12 @@ func TestZipCreateWithGitPath(t *testing.T) {
 		paths := []string{
 			createTestGitPathFile(t, singleByte),
 			createTestGitPathFile(t, multiBytes),
+		}
+
+		expectedMode := os.FileMode(0640)
+		if runtime.GOOS == "windows" {
+			// windows doesn't support the same permissions as Linux
+			expectedMode = 0666
 		}
 
 		f, err := os.Create(fileName)
@@ -182,11 +190,11 @@ func TestZipCreateWithGitPath(t *testing.T) {
 		assert.Len(t, archive.File, 2)
 
 		assert.Equal(t, paths[0], archive.File[0].Name)
-		assert.Equal(t, os.FileMode(0640), archive.File[0].Mode().Perm())
+		assert.Equal(t, expectedMode, archive.File[0].Mode().Perm())
 		assert.NotEmpty(t, archive.File[0].Extra)
 
 		assert.Equal(t, paths[1], archive.File[1].Name)
-		assert.Equal(t, os.FileMode(0640), archive.File[1].Mode().Perm())
+		assert.Equal(t, expectedMode, archive.File[1].Mode().Perm())
 		assert.NotEmpty(t, archive.File[1].Extra)
 	})
 }
