+++ helpers/archives/zip_create_windows_test.go
@@ -1,0 +1,11 @@
+// +build windows
+
+package archives
+
+import (
+	"testing"
+)
+
+func createTestPipe(t *testing.T, csb charsetByte) string {
+	panic("unsupported - this should not be called")
+}
